warning fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@16941 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/IEView/src/IEView.cpp
@@ -1159,7 +1159,7 @@ VARIANT IEView::CallJScript(const CMString strFunc, size_t nPCount, const CMStri
 	
 	DISPPARAMS dispparams;
 	memset(&dispparams, 0, sizeof dispparams);
-	dispparams.cArgs = nPCount;
+	dispparams.cArgs = (UINT)nPCount;
 	dispparams.rgvarg = new VARIANT[dispparams.cArgs];
 	dispparams.cNamedArgs = 0;
